addresses everything except timeoutPromise
