+++ .eslintrc.json
@@ -17,20 +17,29 @@
       "error",
       {
         "zones": [
-          // this means that in the src/www/infrastructure folder,
+          // This means that in the src/infrastructure folder,
           // you can't import any code,
-          // with the exception of other files within the src/www/infrastructure folder
+          // with the exception of other files within the src/infrastructure folder.
           {
             "target": "./src/infrastructure",
             "from": ".",
             "except": ["./src/infrastructure", "./node_modules"]
           },
-          // similar to above but for src/www/model, but you can use files from both the
-          // src/www/model and src/www/infrastructure paths
+          // Similar to above but for src/www/model, but you can use files from both the
+          // src/www/model and src/www/infrastructure paths.
           {
             "target": "./src/www/model",
             "from": ".",
-            "except": ["./src/www/model", "./src/www/infrastructure", "./node_modules"]
+            "except": ["./src/www/model", "./src/infrastructure", "./node_modules"]
+          },
+          // Prevents the internals of the outline_server_repository from being used publicly in the app.
+          {
+            "target": "./src/www/app/*.ts",
+            "from": "./src/www/app/outline_server_repository/server.ts"
+          },
+          {
+            "target": "./src/www/app/*.ts",
+            "from": "./src/www/app/outline_server_repository/access_key_serialization.ts"
           },
           {
             "target": "./src/www/views",
